+++ lib/core/enum/currency_direction.dart
@@ -1,3 +1,3 @@
-part of '../flutter_extend.dart';
+part of '../../flutter_extend.dart';
 
 enum CurrencyDirection { left, right }