+++ lib/files/enum/file_type_enum.dart
@@ -1,0 +1,12 @@
+enum FileType {
+  image,
+  video,
+  audio,
+  document,
+  executable,
+  archive,
+  code,
+  vector,
+  hidden,
+  unknown
+}